evaluator: move "PluginResult" to "tftbase"
We have some simple data classes that don't have additional dependencies
or logic. "PluginResult" is one of them. Move it from "evaluator" to
"tftbase" module. That makes more sense, because next we will add a
"pluginbase" module, that will require the "PluginResult" and be used by
"evaluator". It works bad, if "PluginResult" is inside "evaluator"
module.

To avoid cycles, move the simple types "PluginResult" to "tftbase".
